--- a/server/core/src/main/java/cc/blynk/server/core/model/auth/Session.java
+++ b/server/core/src/main/java/cc/blynk/server/core/model/auth/Session.java
@@ -13,19 +13,20 @@ import io.netty.util.internal.ConcurrentSet;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
+import java.util.HashSet;
 import java.util.Set;
 import java.util.stream.Collectors;
 
-import static cc.blynk.utils.ByteBufUtil.*;
-import static cc.blynk.utils.StateHolderUtil.*;
+import static cc.blynk.utils.ByteBufUtil.makeResponse;
+import static cc.blynk.utils.ByteBufUtil.makeStringMessage;
+import static cc.blynk.utils.StateHolderUtil.getHardState;
 
 /**
  * The Blynk Project.
  * Created by Dmitriy Dumanskiy.
  * Created on 2/1/2015.
- *
+ * <p>
  * DefaultChannelGroup.java too complicated. so doing in simple way for now.
- *
  */
 public class Session {
 
@@ -84,13 +85,13 @@ public class Session {
     }
 
     public boolean sendMessageToHardware(int activeDashId, short cmd, int msgId, String body) {
-        Set<Channel> targetChannels = hardwareChannels.stream().
-                filter(channel -> {
-                            HardwareStateHolder hardwareState = getHardState(channel);
-                            return hardwareState != null && hardwareState.dashId == activeDashId;
-                        }
-                ).
-                collect(Collectors.toSet());
+        Set<Channel> targetChannels = new HashSet<>();
+        for (Channel channel : hardwareChannels) {
+            HardwareStateHolder hardwareState = getHardState(channel);
+            if (hardwareState != null && hardwareState.dashId == activeDashId) {
+                targetChannels.add(channel);
+            }
+        }
 
         int channelsNum = targetChannels.size();
         if (channelsNum == 0)
@@ -151,9 +152,12 @@ public class Session {
     }
 
     public void sendToSharedApps(Channel sendingChannel, String sharedToken, short cmd, int msgId, String body) {
-        Set<Channel> targetChannels = appChannels.stream().
-                filter(appChannel -> appChannel != sendingChannel && needSync(appChannel, sharedToken)).
-                collect(Collectors.toSet());
+        Set<Channel> targetChannels = new HashSet<>();
+        for (Channel channel : appChannels) {
+            if (channel != sendingChannel && needSync(channel, sharedToken)) {
+                targetChannels.add(channel);
+            }
+        }
 
         int channelsNum = targetChannels.size();
         if (channelsNum == 0)