## With lambdas
```bash
Result "sendMessageToHardware_1Channel":
  421.681 ±(99.9%) 10.155 ns/op \[Average\]
  (min, avg, max) = (412.314, 421.681, 454.206), stdev = 9.499
  CI (99.9%): \[411.526, 431.836\] (assumes normal distribution)

Result "sendMessageToHardware_2Channels":
  482.553 ±(99.9%) 9.576 ns/op \[Average\]
  (min, avg, max) = (460.661, 482.553, 497.359), stdev = 8.958
  CI (99.9%): \[472.976, 492.129\] (assumes normal distribution)

Benchmark                                        Mode  Cnt    Score    Error  Units
SessionPerfTest.sendMessageToHardware_1Channel   avgt   15  421.681 ± 10.155  ns/op
SessionPerfTest.sendMessageToHardware_2Channels  avgt   15  482.553 ±  9.576  ns/op

```
## With no lambdas (plain if-s)
```bash
Result "sendMessageToHardware_1Channel":
  344.877 ±(99.9%) 6.639 ns/op \[Average\]
  (min, avg, max) = (336.088, 344.877, 356.549), stdev = 6.210
  CI (99.9%): \[338.238, 351.516\] (assumes normal distribution)

Result "sendMessageToHardware_2Channels":
  410.046 ±(99.9%) 8.168 ns/op \[Average\]
  (min, avg, max) = (397.710, 410.046, 420.379), stdev = 7.641
  CI (99.9%): \[401.877, 418.214\] (assumes normal distribution)

Benchmark                                        Mode  Cnt    Score   Error  Units
SessionPerfTest.sendMessageToHardware_1Channel   avgt   15  344.877 ± 6.639  ns/op
SessionPerfTest.sendMessageToHardware_2Channels  avgt   15  410.046 ± 8.168  ns/op
```