@artem-v 2016-05-12
New  results for 1,2,3,4 channels (been running on another workstation). 
## With no lambdas (plain if-s)
```
Benchmark                                        Mode  Cnt    Score    Error  Units
SessionPerfTest.sendMessageToHardware_1Channel   avgt    5  385.810 ± 34.583  ns/op
SessionPerfTest.sendMessageToHardware_2Channels  avgt    5  476.098 ± 84.936  ns/op
SessionPerfTest.sendMessageToHardware_3Channels  avgt    5  496.396 ± 57.958  ns/op
SessionPerfTest.sendMessageToHardware_4Channels  avgt    5  512.797 ± 83.357  ns/op
```