@doom369 2016-05-12
@artem-v Thanks for benchmark! I really appreciate this. I run your tests. I have similar results. So my initial suggestion was correct. You improved method in general. But for most popular case (1 to 1) proposed solution is a bit slower (20% in my results).